Removed all/most warnings and messages for x64 and x86.
[Add] Window.h, VulkanDebug.h, VulkanInitializers.h, ResourceLoader.h, PipelineData.h, FrameData.h, BufferData.h, Material.h, Renderer.cpp, DescriptorHandler.h.  They now all ignore warnings related to external libraries.

[Change] VulkanRenderer.sln, removed ARM and ARM64 options.
[Change[ VulkanEngine.vcxproj, now ignores the LIBCMT as nodefaultlib:

[Change] GUI.cpp, removed unreferenced local variable io.
[Change] GUI.h, changed VulkanDevice forward declaration from class to struct.
[Change] GUI.h, m_ImguiPool now standard a VK_NULL_HANDLE

[Change] DescriptorHandler.cpp, forloops int to size_t for x86 support.
[Change] DescriptorHandler.cpp, for each now references the variable instead of making a copy.
[Change] DescriptorHandler.h/cpp, size_t for variables changed to uint64_t for x86 support.

[Change] ObjectManager.cpp, UpdateObjects function, now doesn't send the deltatime to Renderer and makes a_Dt a void.
[Change] Renderer.cpp, CreateSwapChain function, now casts the vector.size to a uint32_t.
[Change] Renderer.cpp/h, DrawFrame, now doesn't ask for delta time.
[Change] Renderer.cpp, forloop int now size_t for x86 support.

[Change] SceneObject.h, BaseRenderObject.h, RenderShape.h, CameraObject.h, now all have virtual destructors.

[Change] Transform.h, Transform is now a class instead of a struct.

[Change] RenderFactory.h/cpp, CreateRenderObject function, now doesn't ask for a reference to a bool.

[Change] FrameData.h, all variables are now standard a VK_NULL_HANDLE

[Change] Scene.cpp, RemoveModelID function, now doesn't return false on the end since it was unreachable because of the throw runtime_error.

[Change] SceneManager.h/cpp, CurrentScene is now a int32_t instead of a uint32_t, since it should be possible to be a -1.
[Change] SceneManager.h, Transform and Material are forward decelerated as a class instead of a struct.
[Change] SceneManager.cpp, UpdateScene, now voids the deltatime given.

+++ VulkanRenderer/VulkanRendererWindows/MainRenderer/VulkanSwapChain.cpp
@@ -1,5 +1,8 @@
 #include "VulkanSwapChain.h"
 
+#include <stdexcept>
+#include <algorithm>
+
 VulkanSwapChain::VulkanSwapChain()
 {}
 
@@ -9,7 +12,7 @@ void VulkanSwapChain::ConnectVulkanDevice(VulkanDevice* a_VulkanDevice)
 }
 
 
-void VulkanSwapChain::CreateSwapChain(Window& r_Window, const size_t a_FrameAmount)
+void VulkanSwapChain::CreateSwapChain(Window& r_Window, const uint32_t a_FrameAmount)
 {
 	VkSwapchainKHR t_OldSwapChain = SwapChain;
